refactor(03.2): DRY wishlist summary type + relative-ago helper
- add WishlistSummaryDto to dto.ts as the one home for the per-listing
  summary wire shape ({ balance, lastDate, recentDays }); the four /games
  components stop re-declaring it inline
- extract the verbatim relativeAgo bucketing into src/lib/wishlist-ago.ts
  (wishlistAgo), routed through the same m.wishlist_ago_* messages
- WishlistSummary + StoresSection consume the shared type/helper
